test(Youtube): rework search implementation
The substitute api module now only handles the HTTP request, allowing
the API controller test to cover the parsing of the response.
