--- lib/juicebox/youtube/in_memory.ex
@@ -1,19 +1,0 @@
-defmodule Juicebox.Youtube.InMemory do
-  @behaviour Juicebox.Youtube
-  alias Juicebox.Youtube.Video, as: Video
-
-  def search("bad") do
-    {:error}
-  end
-
-  def search(_) do
-    videos = [%Video{
-      video_id: 1,
-      title: "Movie Bloopers",
-      description: "Some funny movie bloopers",
-      thumbnail: "https://i.ytimg.com/vi/1IZldRSGW50/hqdefault.jpg"
-    }]
-
-    {:ok, videos}
-  end
-end
