+++ web/controllers/api_controller.ex
@@ -1,11 +1,8 @@
 defmodule Juicebox.ApiController do
   use Juicebox.Web, :controller
-  import Juicebox.Youtube
-
-  @youtube_api Application.get_env(:juicebox, :youtube_api)
 
   def videos(conn, params) do
-    case @youtube_api.search(params["q"]) do
+    case Juicebox.Youtube.Client.search(params["q"]) do
       {:ok, videos} ->
         render(conn, videos: videos)
       _ ->
